+++ polaris-pipeline/text-extractor.tests/Functions/RemoveCaseIndexesTests.cs
@@ -132,7 +132,7 @@ namespace text_extractor.tests.Functions
             _errorResult = new JsonResult(_fixture.Create<string>()) { StatusCode = (int)HttpStatusCode.InternalServerError };
             _mockExceptionHandler.Setup(handler => handler.HandleExceptionNew(It.IsAny<Exception>(), It.IsAny<Guid>(), It.IsAny<string>(), _mockLogger.Object))
                 .Returns(_errorResult);
-            _mockSearchIndexService.Setup(s => s.RemoveCaseIndexEntriesAsync(It.IsAny<long>(), It.IsAny<Guid>()))
+            _mockSearchIndexService.Setup(s => s.RemoveCaseIndexEntriesAsync(It.IsAny<int>(), It.IsAny<Guid>()))
                 .ThrowsAsync(new Exception("Test exception"));
 
             var response = await _removeCaseIndexes.Run(mockRequest.Object, _caseId);
