+++ polaris-pipeline/text-extractor/Functions/DocumentIndexCount.cs
@@ -1,5 +1,3 @@
-using System;
-using System.Threading.Tasks;
 using Common.Configuration;
 using Common.Extensions;
 using Common.Handlers;
@@ -30,7 +28,7 @@ namespace text_extractor.Functions
         }
 
         [Function(nameof(DocumentIndexCount))]
-        public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = RestApi.DocumentIndexCount)] HttpRequest request, long caseId, string documentId, long versionId)
+        public async Task<IActionResult> Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = RestApi.DocumentIndexCount)] HttpRequest request, int caseId, string documentId, long versionId)
         {
             var correlationId = Guid.Empty;
